requirements.txt
websockets==11.0.3
PyQt5
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0 
PySide6
qtawesome